Remove  functions and improve README
