+++ Sources/Clamped/BinaryFloatingPoint+Clamped.swift
@@ -1,15 +1,11 @@
 import Foundation
 
- public extension BinaryFloatingPoint {
-     func clamped(to range: ClosedRange<Self>) -> Self {
-         clamped(lowerBound: range.lowerBound, upperBound: range.upperBound)
-     }
-
-     func clamped(to range: Range<Self>) -> Self {
-         clamped(lowerBound: range.lowerBound, upperBound: range.upperBound)
-     }
-
-     func clamped(lowerBound: Self, upperBound: Self) -> Self {
-         min(upperBound, max(lowerBound, self))
-     }
- }
+public extension BinaryFloatingPoint {
+    func clamped(to range: ClosedRange<Self>) -> Self {
+        clamped(lowerBound: range.lowerBound, upperBound: range.upperBound)
+    }
+    
+    func clamped(lowerBound: Self, upperBound: Self) -> Self {
+        min(upperBound, max(lowerBound, self))
+    }
+}
